Moved files to appropriate folders
- for visibility
- faster and easier searching
