+++ src/components/Cards/Requester/CardRequestDocuments.svelte
@@ -19,6 +19,13 @@
                 Document
             </th>
             <th
+                    class="px-6 align-middle border border-solid py-3 text-xs uppercase border-l-0 border-r-0 whitespace-nowrap font-semibold text-left {color ==='light'
+                        ? 'bg-blueGray-50 text-blueGray-500 border-blueGray-100'
+                        : 'bg-red-700 text-red-200 border-red-600'}"
+            >
+                File
+            </th>
+            <th
                     class="px-6 align-middle border border-solid py-3 text-xs uppercase border-l-0 border-r-0 whitespace-nowrap font-semibold text-left {color ===
                       'light'
                         ? 'bg-blueGray-50 text-blueGray-500 border-blueGray-100'
@@ -38,13 +45,25 @@
                 >
                           <span
                                   class="ml-3 font-bold {color === 'light'
-                              ? 'btext-blueGray-600'
+                              ? 'text-blueGray-600'
                               : 'text-whit'}"
                           >
                            { doc.FileName}
 
                           </span>
                 </th>
+                <td
+                        class="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4"
+                >
+                          <span
+                                  class="ml-3 font-bold {color === 'light'
+                              ? 'text-blueGray-600'
+                              : 'text-white'}"
+                          >
+                           { doc.FilePurpose}
+
+                          </span>
+                </td>
                 <td
                         class="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4"
                 >
