+++ src/components/Cards/Requester/CardUserRequesterTable.svelte
@@ -1,8 +1,8 @@
 <!-- CardApprovalsTable.svelte -->
 <script>
   import {onMount} from "svelte";
-  import UserRequestTable from "./Requester/UserRequestTable.svelte";
-  import {auth} from "../../authentication/AuthStore";
+  import UserRequestTable from "./UserRequestTable.svelte";
+  import {auth} from "../../../authentication/AuthStore";
   import {navigate} from "svelte-routing";
 
   const env = process.env.config;
